server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1
bcrypt==4.1.3
PyJWT==2.8.0
slowapi==0.1.9
python-multipart==0.0.9
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.7
